Separate method into multiple classes for lisibility
