+++ KruacentE.Misc/ServerHandler.cs
@@ -14,13 +14,13 @@ namespace KE.Misc
             if(MainPlugin.Instance.Config.ChanceFF >= 0 && MainPlugin.Instance.Config.ChanceFF <= 100)
                 MainPlugin.Instance.RandomFF();
             if (MainPlugin.Instance.Config.ChanceClassDDoorGoesBoom >= 0 && MainPlugin.Instance.Config.ChanceClassDDoorGoesBoom <= 100)
-                MainPlugin.Instance.ClassDDoorGoesBoom();
+                MainPlugin.Instance.ClassDDoor.ClassDDoorGoesBoom();
             if(MainPlugin.Instance.Config.AutoNukeAnnoucement)
                 Timing.RunCoroutine(MainPlugin.Instance.NukeAnnouncement());
             if(MainPlugin.Instance.Config.PeanutLockDown)
                 Timing.RunCoroutine(MainPlugin.Instance.PeanutLockdown());
             if(MainPlugin.Instance.Config.AutoElevator)
-                Timing.RunCoroutine(MainPlugin.Instance.AutoElevator());
+                Timing.RunCoroutine(MainPlugin.Instance.AutoElevator.StartElevator());
         }
     }
 }
